fix: unbreak Android 14 exports and close release-gate holes
ExportService declared only the API 35+ mediaProcessing FGS type, so on
Android 14 startForeground had no recognized type and every export
crashed at start — declare dataSync for API <= 34 and start typed on Q+.
Target-size exports floored video bitrate at 500 kbps with untouched
audio, producing 2x the promised file size on longer timelines; audio
now steps down toward 64 kbps first and the floor drops to 250 kbps.
Clip duration math divided Long millis by Float, losing ms precision
past ~4.7 h. Release verification regains the public-claims and audio-
API-policy gates orphaned by the GitHub Actions removal (and the README
claims they caught), pins the release signer to the certificate every
shipped GitHub Release actually carries (sidecar freshness alone cannot
detect a wrong key), fails closed on unparseable native libraries in the
16 KB alignment gate, and drops dead workflow-era remnants.

+++ app/src/main/java/com/novacut/editor/model/Project.kt
@@ -233,10 +233,13 @@ data class Clip(
             return if (timelineDurationMs.isFinite() && timelineDurationMs > 0.0) {
                 timelineDurationMs.toLong()
             } else {
-                (trimRange / finitePositiveSpeed(speed)).toLong()
+                // Divide in Double: Long / Float promotes to Float, whose
+                // 24-bit mantissa loses millisecond precision past ~4.6 h and
+                // drifts against the Double-based offset mappers below.
+                (trimRange / finitePositiveSpeed(speed).toDouble()).toLong()
             }
         }
-        return (trimRange / finitePositiveSpeed(speed)).toLong()
+        return (trimRange / finitePositiveSpeed(speed).toDouble()).toLong()
     }
     val timelineEndMs: Long get() = timelineStartMs + durationMs
